requirements.txt
# general app
django >3.2,<3.3
python-decouple
pillow
djangorestframework



# third party modules
markdown       
django-filter  
drf-yasg[validation]
djangorestframework-simplejwt
djoser

#reformating and tests
black
flake8



# background proccess and cache
celery